+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 *Other*
 - Kubernetes 1.9 is no longer officially supported as of this version
 
+*Bug fixes*
+- Attempting to deploy from a directory that only contains `secrets.ejson` will no longer fail deploy ([#416](https://github.com/Shopify/kubernetes-deploy/pull/416))
+
 ## 0.23.0
 
 *Features*