+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## next
 
+*Enhancements*
+- Uses `TASK_ID` environment variable as the `deployment_id` when rendering resource templates for better Shipit integration.
+
 *Bug fixes*
 - Attempting to deploy from a directory that only contains `secrets.ejson` will no longer fail deploy ([#416](https://github.com/Shopify/kubernetes-deploy/pull/416))