+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## next
 
 *Enhancements*
-- Uses `TASK_ID` environment variable as the `deployment_id` when rendering resource templates for better Shipit integration.
+- Uses `TASK_ID` environment variable as the `deployment_id` when rendering resource templates for better [Shipit](https://github.com/Shopify/shipit) integration. ([#430](https://github.com/Shopify/kubernetes-deploy/pull/430))
 
 *Bug fixes*
 - Attempting to deploy from a directory that only contains `secrets.ejson` will no longer fail deploy ([#416](https://github.com/Shopify/kubernetes-deploy/pull/416))