+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 0.25.0
 
 *Bug fixes*
 - Attempting to deploy from a directory that only contains `secrets.ejson` will no longer fail deploy ([#416](https://github.com/Shopify/kubernetes-deploy/pull/416))